Add files via upload
"This project is a dynamic web application that utilizes a MySQL database to store user registration information. The database contains a 'person' table which serves as the primary storage for all registered users. The registration process involves filling in a form with personal details such as name, email, and password. This information is then securely stored in the 'person' table, allowing users to access their account and update their information as needed. The database is designed to be scalable and efficient, making it easy to manage a large number of users and their information. With its user-friendly interface and robust database, this project is a powerful tool for managing user registrations and personal data."